Add skeleton documentation of api
The autoapi extension of sphinx is used to setup doc generation. As
the codebase builds, automated doc generation should really help. It
still needs a lot of tweaking and tutorials need to be added, but
these can come in later.

In the future, the plan is to use github actions to automate doc
generation on push and host them on gh pages.