Add initial skeleton for napi
The skeleton depicts the plan for napi. There are three necessary
functionalities expected from napi - raw data format convertion,
indexing the data for quick querying, and generation of reports based
on the index.

We envision the first release of napi to have two submodules - data
and analysis. data will enable the functions mentioned above in a
hierarchical manner. A file tree forms a layout and a collection of
layouts forms a dataset. The specification followed is defined as
YAML.

This might change as the implementation takes shape.