Add path building function to Specification
Provided the entities, the path_build() function can construct paths
by refering to the yaml specification file. The yaml specification
file is a private attribute of dict type in the class. Further,
entities can optional or mandatory with pre-defined default and valid
values.

A guide on writing the yaml specification file is necessary and should
be done at some point.