Add organization functionality
There might be cases were the files are in the right format, but are
unorganized and not named according to the specification. In this
case, if identifiers are present in the file name, this function
enables organizing them into a format that abides by the
specification.

This is made possible by user-defined rules. These provide the
function with information regarding feature extraction. Further,
fellows who share the directory space with an eligible data file can
be renamed and organized too.