Fix bug in string join
String join method expects an iterable as argument. The arguments were
provided before not in this manner and this raised an error. This was
not noticed before as the flags that use these in the rules config
were not tested before. Fixed this by enclosing the arguments as a
list.