Add sqlalchemy as dependency
Once data is organized according to a specification, we would like to
query files given filters and maybe generate quick reports. Indexing
the dataset and storing the component files would enable these. ORM
design by using sqlalchemy allows quick development of this.

Broad idea is that a dataset can consist of multiple modalities or
layouts. Layouts are in turn made up of many files and
directories. Each file or directory can have a number of tags
associated to it. These tags are defined by the specification.