Add initial db interface
To connect to the database for inserting info and querying, an
interface is required. This module will provide the functionalities
expected from an interface.

sqlite is chosen as the database will primarily be used to store the
indexed dataset information which will not place a huge load. A
production environment such as mariadb or postgresql would be
overkill.

Bug note: Initial attempts to return session directly from
sessionmaker as `return sessionmaker()` did not reflect changes in db
file despite commit. Instead of returning a different session object
each time, storing a copy of a single session and returning that in
future calls resolved this.