Add basic file indexing
Once a layout is initialised, all files in it are indexed using an
Indexer in an iterative manner. Both directories and files are
indexed. Next step would be to index tags defined in specification.

Bug note: Circular imports were an issue as `Layout` needs `File` and
`File` in turn needs `Layout`. To circumvent this, `File` was moved
along with other components in files.py to layout.py

Todo list:
  - Using back_populates relationship to reflect changes to layout in
  files. Will this be needed? Ideally layout properties should not
  be changeable by user or be changes.
  - Returning only relative path with respect to layout root to user
  instead of whole path.