Add yaml reading to utils
napi heavily relies on configs written in yaml for organization and
specification rules. Soon, it will depend on yaml files for file
conversion to other formats. Thus, it would be good to have yaml
reading capability.

At a later point, there can be a config manager so that all yaml
config files can be in a single file and managed in one place. But
without it, napi should still be functionable.