Enhance readability by renaming imports and entity to tag
We use dictionary `copy()` to create a copy of the tags of the data
file. This makes sure that we do not change the original set of tags
attached to the file while either trying to match the appropriate path
pattern or creating tags for fellow files. But this conflicts with the
`copy()` defined in utils. Importing utils circumvents this. Instead,
we use `utils.copy()`.

As per our specification design, a file can have many tags attached to
it. To reflect this better, renaming entity to tag in standard.py.