Add default specification
It would be nice to have a default specification if none is
specified. This allows a certain constraint on the type of files in
the dataset based on path patterns allowed. This can always be
overcome by providing an appropriate dict hack. Further, it would be
nice to allow handling of both paths to yaml files or dicts. This
commit allows these.

While at this, unnecessary underscore(read private) variables were
trimmed to non-underscore(real public) variables.

Also, note the change in order for `Specification` parameters. This is
because the spec is more important than providing a name for the
specification.