Remove redundant private variables and make
To enhance readability redundant underscore(read private) variables in
dataset.py are trimmed.

Bug note: In this process, _layout_names was changed to
layout_names. But this introduced a bug as both property and class
attribute now had the same name. Due to this, a RecursionError was
thrown as maximum depth was reached. Property was removed and replaced
with only class attribute to prevent this.