Add tags to indexing
All files in a layout are indexed. We would like the tags associated
to be indexed too. These can be extracted from the specification
associated with a layout. Function for this extraction was added in
commit 3e0856b0fa3263c3d46e8394490794443707d9c5.

These tags are stored in 'tags' table in the sqlite db. This should
help in report generation and more in-depth querying.