Add reporting to datasets or layout
This commit adds the capability to generate high level reports on
a dataset or layout. Simple messages regarding the layouts in the
dataset and their subject numbers are printed.

The idea is to expand this to provide lengthy wordy reports if
needed. Later, exporting it to html or other formats is also an options.