Display relative path instead of absolute for File
Revealing the complete path of a file in not needed. Relative path are
sufficient. While this does not greatly enhance security, it helps
give the idea that all data is in the same directory even though they
might not.

It is possible to have a dataset consisting of different layouts, each
in a different directory with no common root in near levels.