Add querying functionality to Layout
The get_files() query the indexed db for files that match the tag
requirements and returns them. Only files that fit all tag criteria
are returned.

For get_files() to work on already an already saved index, we can
follow the below steps:
  - create SessionManager for index db
  - create Indexer that links layout to index db
  - Pass this as an argument to Layout with arg index_layout=False
The index_layout argument is false as we are using an already indexed
layout. Later maybe, 'force' can be added.

The query functionality should be fast as it completely translated the
filters to sql instead of a partial python-sql mix. The get_files() in
turn calls get() in Indexer which will act as the gateway of sorts to
query from indexed db. At a later point, alternatives to all data or
lazy-loading can be added.