Add relationship between ORM objects
In addition to specifying the relationship between objects using
foreign keys in the sql table, mentioning the relationship as a Mapped
object between the ORM classes is useful in better querying and
access. Thus, the relationship has been added. This in turn also leads
to cleaner and more readble code.

Minor changes to the indexing was added to take advantage of
this. Instead of adding each object individually, addition of a single
linked objects leads to addition of other relevant objects too.