Add better ORM object relationships
It would be ideal if an attribute of files is able to access all its
tags. And each tag has a link back to its file. Similarly, the
relationship between a layout and its files. While this was present
before in a different form, it was possible to make it close the
actual relationship and in the process faster.

Instead of a list representing all files in a layout, it is now
represented by a set. This makes it easier to check if a file is
already part of a layout during indexing, hence makes the future work
of updates only indexing easier.

The collection of tags linked to a file is now a dictionary. This
leads to a more readable accessing of a tag and also verifying if a
tag is present for a file using get(). This is cleaner.