Add robust tag reporting
It is not a good idea to depend on a strict tag name, previously
subject to generate the report. It is possible that a different
dataset project used different a naming scheme. Hence, it should be
possible to mention the tag names that should be considered during
report generation. This commit adds this ability. The target tags for
counting are provided as a list.