Add sourcedata to bids specification
The sourcedata directory inside the bids directory store the raw
source data. The clean and curated data is generated from this. It
would be nice if this can be indexed too. This can in turn help in
using the indexing functionality for querying files to be
converted. This addition is made in this commit.