Add logging to organize functionality
Logging makes life easier during debugging and to also trace things
back at a later point. It is essential that it is present. This is
added with this commit.

Some rule values such as source and destination are stored in variable
instead being accessed when needed using get(). This might lead to a
minor non-noticable speed-up, but the main reason this was done was so
that these values can be used in logs.

Explicit checking of whether the output path is build is made unlike
before. Again to log this scenario.

Bug note: The value rule is overriding in all cases. While using
continue made the flow easier to read in my opinion, it lead to the
tag logs during its use not being added to the log buffer. To overcome
this, we have switched to if else instead.