Rename run_command function to run_shell
This change was made with the idea of having another function
run_matlab that can matlab commands. In that case, one having the name
command and the other not is not ideal. So, this was changed. Based on
the utility of matlab toolkits that python tools do not provide,
run_matlab might be added at a later point. None the less, this
renaming was needed.