Fix bug that caused directory files to not be organized
Organization using rules worked for all modalities that had a file as
the target data content. Despite having rules that matched the pattern
and the pattern verified online through pythex, the organization did
not function as expected for target data content that were
directories. For example, organizing of target pattern '*.mff' did not
work for EEG as usually mff (EGI file format) is a directory with more
files inside it.

This issue was cause by two things:

1) In the function that gets matches to the pattern, only files were
checked for a match and not directories. This was rectified by add
direcotries to the list of contents that were checked for a match.

2) In the function that copies the matched data content to the
destination, copy2 was used. This only supported copying files. This
was fixed by using copytree in case of directory targets and copy2 in
case of file targets. A simple search did not yield a common fucntion
that worked for both types.

Minor changes were made to accomodate these:

  - The matching function was renamed from get_matching_files() to
  get_matches() to include matching directories too.
  - The copy() docstring was changed to reflect copying of directories
  too.