Add EEG format conversion functionality
In most cases the data files are not in the format we need or is
recommended by some specification and some form of format conversion
is needed. Instead of reimplementing this, it would be nice if we
could take advantage of existing tools to do this. To this extent,
this commit does this for EEG using the mne toolkit. Extensive logging
is added for prevenance tracking. It can be configured for multiple
formats as required and has anonymization added in too.