Add validate functionality to Specification
It is important to have a way to validate if a path abides the
specification defined. This can be implemented by checking if the path
reconstructed from tags retrieved matches the provided relative path
from root. On match, the provided path is valid. This is implemented
and will be added in this commit.