Add CI workflow for docs generation
It would be ideal if documentation is generated automatically on every
push and is available on a site. It is common to use either Sphinx or
MKDocs + Git Pages or ReadThedocs for this purpose. We have tried to
use Sphinx to generate documentation and host it on Git
Pages. ReadThedocs is great too!

A workflow builds HTML documentation on push to main and this is added
to the gh-pages brach for hosting. A requirements file is needed as we
use an external theme rtd-theme.