Update poetry dependencies
Downgraded python version from 3.10 to 3.9 for more stability and as
some modules did not support 3.10.

Manually added seaborn as a dependency as mne-nirs depends on it but
does have it mentioned in its dependency list causing it to raise an
error during import time. This temporarily fixes it.

Added libraries to work with data. These will come in handly during
the conversion functionality release.