Add ability overwrite files and directories
Previously, copy will throw an error if the file already exists. This
does not allow overwriting files which might be needed during certain
organization rules or strategies. For example, if both
sub-2/eyetrack.asc and sub-134/EO/eyetrack.asc are data files. Then,
a possible strategy is to match for sub-\d+|EO and overwrite sub with
EO once found to level the hierarchy. To enable situations like this,
overwrite has been introduced.

Overwrite is a two step process - first remove directory or file if
exists, then write to same location. Different functions are used
depending on whether the target is a file or a directory.