Add log if more than two values found for tag
Ideally, we would like all directories to be organized to have unique
tags. But it might happen that, since they are disordred they have
multiple matches for a single tag. For example,

  - sub-140/scan.txt
  - sub-140/sub-150/scan.txt

In this case, since a subject is nested in another, the second path
has two matches for subject tag - namely 140 and 150. Though in most
cases considering the last match should resolve the problem, sometimes
it does not work and manual inspection is needed. This commit adds a
log entry for these occurance for later inspection.