Add ability to add file to all organized files
There might arise a need where a file has to be added to all organized
files. Either since it is absent in all but needed or in partially
present in a subset of the organized files. For example, say there
montage information is absent in all organized EEG files but is
uniform and this has to be added during organization. This commit
enables this.

The path of the file that is to be added and its positon relative to
the organized file are required in rules. Position can take two values
- content if it is to be nested inside the organized file, which is
possible only if the organized file is a directory or fellow if it is
to be alongside the organized file.