Add prefix and kind attributes to Layout
There are different types of layout: primary, sourcedata or
derivative. Each have their place in the specification. Since the
specification should be relative to root of the dataset, sourcedata
and derivatives need a prefix before their path as their root in
nested inside the dataset's root. For example, the root of the dataset
might be /data/bids while the root of the sourcedata will be
/data/bids/sourcedata/eeg. In these cases, sourcedata/eeg has to be
prefixed to files from this layout. To implement this, two attributes
named prefix and kind are added to the Layout.

These are decipher automatically based on the root name and are
initiliased on object creation - both by sqlalchemy using
reconstructor or by python initialize using the constructor. To make
this possible, it is now recommended to create a layout using the
static factory method create() that has been added to Layout. create()
will return the suitable Layout with appropriate attributes set.

Other approaches tried during implementation included attempts to
modify __new__() that is called during object creation. But this was
not elegant and an object factory was the best approach.