Add ability to stop indexing at valid directories
Indexing huge directories such as DICOM sourcedata where only the
directory is valid but the content images do not pass the
specification takes a lot of time. It would be ideal to skip contents
if a directory is valid and has been indexed. If not, then the
contents can be explored.

To enable this, file indexing returns a boolean value based on
indexing status. True on success and False on failure. This is used in
combination with validity attribute to decide. Further, the file
indexing now adds the prefix to generate more approriate tags and
prevent skipping files in non-primary layouts.

Note that Indexer attribute `only_valid` has been renamed to
`valid_only` as it more readable. Docstrings have been modified to
more appropriately reflect these additions.