Add ability to convert one format to another
In almost all sourcedata types, it is necessary to convert into
another format before it abides to a specification. Like, .edf ->
.asc, DICOM -> NIfTI, and so on. This module provides the ability to
perform these. Given the files, required output format, and the
destination, the converter does all required job. Conversion specific
parameters such as line_freq and anonymization info can be
provided. Detailed error and log messages are present to guide the
user and also keep track of the process.

The converters depend on external packages and toolboxes in most
cases. Some are added to the dependency list.