Add ability to skip files during matches
There might be cases where certain files that have been flagged have
to skipped from being organized and added to the raw data. This might
be because the session the data belongs to is not resolvable or the
data does not belong to the project or was a test trial. A regex
pattern of this can be provided with the rule using 'skip' keyword.