Add listing of all values possible of a given tag name
It cases where it is not documented what the tag values are, it would
be helpful to have a direct way to list all possible values.  For
example, provided the tag name subject, list all subjects. A method
get_tag_values() is added to Layout for this.

get_tag_values() can also accomodate filters as it uses the same
queries as get_files() an builds on it to retrieve tags. To not
duplicate code, all queries have been moved to Indexer. Queries are
involded and used to retrieve in the Layout.