Fix bugs in shell converts due to bad command formatting
In convert that use shell commands, the extra flags provided are added
to the command template before command building. But due to a bug, all
valid flags were added and this lead to errors. Further, the command
template was formatted to include the parameters but this was not
assigned back to the command, again resulting in errors. In some cases
where the directory did not exist, the command failed. This can be
prevented by creating destination directories before running. These
have been fixed.