Add ability to add tags to file based return value of functions
In certain cases, we might want to pass a file to function and the
store the result of the return as a tag. This can be helpful in
filtering at a later time as this tag is stored. Currently, the
addition supports only boolean functions but we can expands this to
support other return types.