Fix bug due to Layout association with multiple db sessions
By default, each Layout is associated to an Indexer instance. This is
to ensure persistent storage of indexes and to check if a Layout has
been pre-indexed. This default functionality can be changed. During
Layout creation, if the Layout exists in the database then it is
retrieved else a new copy is created. To this copy, an Indexer is
associated. Both of these operations required an Indexer. In cases
where an indexer is not provided for Layout creation, a default
indexer is used.

Using different instances of Indexer for the two operations leads to
the Layout being associated with the sessions of both Indexers. This
is because SQLalchemy merge by default add the object to the
session. This leads to errors. This has been fixed by ensuring the
same Indexer is used for both the operations.