Remove multi-type dict or path arg from get_spec()
A specification can be defined as a dictionary. This dictionary can
be stored in a file as JSON or YAML and retrieved if needed. The
previous method accounted for this by accepting both dict and
string. If string, it is considered a path and the appropriate file is
retrieved. The naming of the specification if not provided is inferred
from the filename. This is not possible in dict. As a temporary fix,
dict support is removed.

Bug note: In the previous implementation, if file is not present, the
non-existent file path is taken as the specification instead of
raising an error.