Repurpose SessionManager to DB-wide manager
SessionManager is used by ORM during indexing for inserting
indexes. These indexes enable easy querying. We would like to expand
this for db to support not just ORM-related indexing but also querying
tables, or creating additional tables in the database as the db might
have other records that are part of the dataset. For this,
SessionManager is purposed to DBManager.

DBManager will support connecting to any db via SQLalchemy db
URLs. All metadata such as tables and columns can be accessed using
the meta attribute of DBManager. This doesnot break indexing
functionality but hopes to add db interfacing on top of it.

For transition to DBManager from SessionManager, all modules that
reference SessionManager are changed to DBManager and its attributes.