Add ability to insert records to table in db
With the ability to insert records, DBManager is equipped to do all
basic operations on a database. The to_table is used to insert into a
table in the database. It provided functionality to check for
duplicates, foreign key presence in parent, and resolution strategies
for each of them. This is implemented as a wrapped around the pandas
to_sql() method.