Add ability to migrate tables from one db to another
A scenario in db management is when the old db needs to be phased out
and new one put in its place. Or if the tables were not defined
appropriately in the first instance, redefine it and move records to
new format. These migration schemes are supported with
migrate(). Given the source, target, and the mapping from source to
target, migrate() moves records based on mapping rules. It creates new
tables if required and flags bad records.