Rename napi/standard.py to napi/specification.py
To more appropriate reflect the functionality, standard is renamed to
specification. The initial idea was for this module to also provide
quality control and generate quality metrics, which has now been
dropped. At a later point, if time permits these can be added in.

To accomodate this renaming, module that depend on this have been
renamed too.