Explicitly mention public objects of modules
It would be nice if often used components in a module can be import
directly from the top level of napi. For example,

    >>> from napi import Specification

instead of,

    >>> from napi.specification import Specification

For this purpose, some components are added as public objects by
explicitly mentioning them in `__all__`. It also states what objects
are imported when `import *` is used.