Add sphinx docs dependencies
The auto-generated documentation will be rewritten appropriately. As a
first step towards this, the appropriate dependencies are added in to
pyproject.toml as well as requirements.txt that is used by Github
Actions.

    - autodoc : Reference docstrings from codebase.
    - napolean : Interpret Numpy style docstrings.
    - viewcode : Directly move to code from API reference.
    - copybutton : Adds copy mechanism to code blocks in docs.

autobuild is only added to poetry as this is used only during
development to view changes live in browser without the need to
rebuilt manually.