Add code of conduct for community to docs
The code of conduct is an important article that describes the
responsibilities of the community. It is even more so important in the
context of sensitive datasets and their usage. We hope this provides
the intention of the spirit expected from the community and that they
are held to the highest standards.