Fix bug for ORM merge when no table present
The SQLalchemy session.merge() expects that the table for the
respective object is present. But since the indexer creates it only
during the first commit which happens in the later stages, an error is
thrown. To prevent this, all tables needed for indexing are created on
initialization of Indexer if absent. This sees the merge to
completion.