Prevent reindexing of layout
Each time a Layout instance is created, the layout is indexed. This
takes significant time if there are a lot of contents. To prevent
this, napi by default does not index if the layout is found in
presistent storage. To force reindexing in case of file changes since
last indexing, the reindex flag can be set to True during call.