Fill in tutorial doc
It is important to have a hold-your-hand interactive tutorial that
introduces napi to a user. The tutorial mimics a typical dataset
workflow. It start with deciding on a Specification, then
organization, and then on to querying which will help with
analysis. In the future, we might want to add more tutorials on db
migration and interfacing with DataLad.