Fix EDF and edf format conflict in convert functionality
For file format conversion functionality, file objects are passed with
the output format mentioned. Conversion proceeds if all file objects
have the same supported extension and if the conversion converter is
found. The supported conversions and their converters were stored in a
dictionary with input file type format as key and the converter
retrieved using this. This led to problems due to edf being a
extension for both EEG and Eye tracking datatypes even they referred
to different formats, namely European Data Format and Eyelink Data
Format. Obviously, since we selected the first converter found and
this led to EEG converter being used on Eyetracking data we
encountered an error.

This has been fixed by using out file type formats supported as
keys. This leads to outputs as expected.