Circumvent process killed cases by removing daysback validation
For EEG file format conversions, once the arguments are parsed and
massaged in the required formats, all files provided were opened in
memory and their raw objects stored as a list. This was needed as all
files metadata needs to be read to get the recording date, based on
which the daysback is validated by comparing the daysback value to
BIDS recommendations like should be before year 1925. Since all files
are open and some part of data loaded, we run out of memory in case
there are a lot of files and the process is killed as a consequence.

This has been circumvented by removing dayback validation. Now, files
are read sequentially and converted one at a time without daysback
validation. Until a better approach to daysback validation is
implemented, this will be the case.

Note: Supported out format EDF has a lower bound on the recording date
it can support. The difference between that and the BIDS recommended
upper bound after daysback is too low to support longtitudinal
multi-year studies.