Fix placeholder mismatch during command formatting in dcm2nii
For conversions that are based on a tool run using a command, we have
a template command that is formatted based on input provided. It was
implemented as individual format steps for each parameter. Apparently,
string format does not work this way and attempts to fill in all
parameters at once and thus raising errors. This has been fixed by
refactoring the code to condense the multiple steps into a single
format line.