Format code to shorten log messages and variables
Some line in the code base are too long and lead to less
readability. To this end, some log messages and variables in certain
cases have been shortened to convey a similar message. There has been
a trade-off as less info on the error is provided but we believe the
content is almost the same and will still aid in debugging.