Add ability to take mapped value from csv while organizing
In certain cases, an identifier in file name can help get another tag
value that we would like to add but is not present in file name. For
example, from a file say a sample tag value 112 is extracted but we
would actually like another id that is not in file name but is
available in a csv that holds a one-to-one mapping from sample ->
id. Then, with this commit it is possible to mention in the config
that we have to consider this mapping to get the required value.

The rule to make this work has been added to docs under the relevant
config.