Refactor db transfer config to better depict case transform
The config for db records transfer allows transforming the case to
either lower or upper. This was provided using the 'transform' key as
the initial idea was to expand this to other possibilities. Since we
do not plan to do this in the immediate future and this is ambiguous,
the key 'case' will be used for this purpose now.