Add logging for debugging transformation of records
Currently, all records that faced an error during transformation are
logged at once making it difficult to narrow down the column that
failed the transformation and debug. Logging individual column
elements that failed transformation would be more helpful for this
purpose. This is added at the cost of a small memory overhead.