Refactor code for transform error to prevent re-calculation
Records that throw error during transformation are tracked in the
transform() method, thus it would be better if we reuse this instead
of recalculating them post-return. To this effect, the code is
refactored such that transform() returns the error along with records,
and this error is used to update overall tr_error.