Add column-wise logging for elements that fail validation
While there exists global logging of records on validation failure, it
is difficult to pinpoint the column that lead to the whole record
failing. This is leads to difficulties debugging and curating the
database records. As a solution, this commit adds logging of elements
in the column that fail validation.