Add ability to ignore insertion of existing records
Attempts to insert an already existing record throws an error if one
of columns is a primary key as they have to be unique. To circumvent
this problem, the ideal solution would be to have a SQL database
invariant 'INSERT IGNORE' query. But since SQLalchemy is yet to have
this, we have resorted to a merge-based solution where the records are
compared with existing records in table, and common records are
removed before insertion.