Refactor create_table() with better args
create_table() currently takes a description, which is the mapping
config completely. Creating the description in one go is tedious and
furthur does not give a clear picture of the args actually used by
this method. To paint a better picture, description is broken down
into parts - the table, columns in the table, and references made by
the table.