Add ability reshape source table
During migration, there might be cases where the new schema is a
normalised version of the older schema. In these cases, it is
required that we are massage the old records into the new schema
without dataloss. Reshape serves this purpose.

Reshaping instructions can be provided under the 'reshape' key by
providing a supported reshaping procedure as key and a sequence of its
args under it. This commit adds 4 initial reshape procedures:

  - adding a column with value
  - splitting a column into smaller multiple columns
  - melting a set of columns into a single column
  - pivoting to wide-form by grouping similar id rows

These are implemented using pandas methods and will thus use the same
arguments for ease of usability.