Fix bug in table creation introduced by reshape functionality
For table creation, the sequence under 'cols' key is used for building
columns. Currently, all columns in 'cols' need to have a name and map
to the source table. And all these columns are created with the new
name. But it is not necessary that all columns are present after
reshaping, for example it is possible that two columns have been
melted into one or a new column have been added. These are very much
possible scenarios.

It should be possible to infer columns that will be lost to reshape
and columns that will be added post-reshape. The most simple way would
be to use a dummy dataframe with all element combinations of each
column and pass through a trial reshape to infer the column
changes. But for now, this commit introduces two new keys to the
mapping config - 'detach' and 'attach'. As the name implies, they
mention columns that will be detached and later new columns that will
be attached.

Now, there are three keys under which columns can be mentioned:

  - cols: For columns that will stay the same post-reshape
  - attach: For column that will added post-reshaping
  - detach: For columns that will be removed in the process of
  reshaping

It is not necessary for attach columns to have a map value but name is
mandatory.