Add datalad as a dependency
It is a requirement to share datasets with collaborators or people
outside the institution. Datalad is a nice tool that enables a
git-like workflow for large data. Data is stored in git-like
repositories based on git-annex and this git-like workflow enables
version control of data. This is a very useful feature. Further,
cloning gives only empty files until explicitly asked. This way
checking out the dataset structure is easy and only the required files
can be retrieved as needed. napi will have methods built on top of
datalad and take advantage of its amazing features.