Add get() to retrieve record as ORM class given primary key
Currently, merge is in use for retreving records that might already
exist in the db. But in cases where any of the attributes has been
already accessed or modified, that is overwritten during merge which
is not desired. So, get() is introduced as an alternative and merge
will be phased out.