Refactor Tag.boolean_tag() to File.add_tag_from_func()
boolean_tag() is used to add a tag to a file based on a function
return value. This is not aptly reflected in its name and its position
under File would be better, since it adds a tag to a file. To this
effect, Tag.boolean_tag() have been refactored into
File.add_tag_from_func(). The return value of the function does not
have to be boolean.