Remove redundant name attribute from Layout
The Layout class maps to a table called 'layouts'. Initially, it was
thought that have a short abbreviation to a long root path would be
helpful and this can be used as an alternative to root in
places. Since, this was not implemented name is redundant as of
now. So, this is removed.