Add ability to get content from remote url
Given an url is setup for a Layout, it is desirable to retrieve
contents from this url when needed or if extra things are added
here. This commit adds this functionality to clone datasets or a file
from url. Currently only urls that are datalad supported are supported
by napi.