Change project name from napi to almirah
It is time for the first release! While it will be filled with bugs,
it is still an occasion to rejoice. But napi unfortunately is taken up
on pypi, so we have chosen to go with almirah instead (datadrobe gave
tough competition). To accomodate this name change, napi had to
replaced with almirah in a lot of places. This commit is a dirty
squish of all of them.