Add release workflow triggered by a tagged push
The typical workflow to release almirah project is like:

  - Bump poetry version and reflect in conf.py
  - Generate changelog with towncrier
  - Tag as X.Y.Z
  - Raise a pull request!

Once merged with main, the release needs to be added to both pypi and
to github release. An automated workflow to do both of these would
make life easier. This commit adds this as a github action that is
triggered when ever a tag is pushed.