Fix bug that lead to path build failure on extra tag presence
The file path is build based on the tags that accompany the file. In
case there are additional tags, namely tags other than the ones
mentioned in the specification associated with a file, path building
fails. This is not desired. Instead, the ideal functionality would be
to build path with the tags necessary and ignore the extra tags. This
is implemented in this commit. The previous functionality can be
brought in if necessary by setting the strict flag to True.