Provide info on data storage architecture setup
For future reference in case the need arises to replicate a similar
data storage infrastructure for a different project, a detailed
document of the current setup will help. For this, the setup info is
added under the ADBS dataset section. The ADBS dataset doc landing
page has been updated to include this.