Update the ADBS dataset reference
The reference is intended to act as THE page to come to if you
interested in getting an overview of the different modalities in the
dataset and their specifics. As a start to populate this, the
necessary broad info about the ADBS project has been added. The
document will be update to add modality specific information based on
the release data.