Refactor replacement of tag value from a map while organizing
Commit 6d056bb7b0e1fb8800b30ac1c99a9b490a84a532 introduced the ability
to replace tag values with value in a mapping. The config key words
used to describe this rule was not very easy to read and it did not
enable the usage of multiple maps for each tag. To remedy this, this
commit simplifies it to a phrase like 'replace col {col_name} with
{replace_col_name} from the file {file_name}'. Further, this can be
added to each tag.