Better log messages to provide more info
Recently, I received suggestions from putative users that it would be
more helpful if the log messages are self-sufficient instead of having
to refer to previous message to get the context first. While
implementing this library-wide will take considerable time and
thought, as a first pass some log messages are updated for this
purpose.