Add padding rule to organization
While organizing datasets, a need may arise to pad certain values with
zero. This may be so that the directories are appropriately
ordered. For example, consider two directories D1, D9, and
D10. Without padding, they will be in the order D1, D10, D9 as by
default ordering is alphabetically in most file managers. But this is
not desired. On padding with length 2, it will be D01, D09, D10, which
is the desired order.

A note of caution though. While padding can be handy, it might mess
with querying. For example, on padding with 5 and querying for files
with subject = '5', no result might be returned despite there being
files belonging to the subject. This is because now the subject is
'00005'. Changes to the querying functionality are need to take this
into account. This has to be added in, but is left out of this commit
for now.